+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Ensure `ElementsUtil.getOverriddenMethod(...)` is public.
+
 ### [v0.47](https://github.com/realityforge/proton/tree/v0.47) (2020-05-19) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.46...v0.47)
 
 Changes in this release: